+++ Content.Client/Atmos/Components/ScrubberVisualsComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Client.Atmos.Components;
+
+[RegisterComponent]
+public sealed class ScrubberVisualsComponent : Component
+{
+}
